feat: add student and teacher attendance management views
- Implemented Attendance.vue for students to view their attendance records with filtering options.
- Created Attendance.vue for teachers to manage attendance, including bulk status updates and attendance recording.
- Added API integration for fetching attendance records and section enrollments.
- Enhanced UI with loading states, error handling, and summary statistics for attendance.

chore: initialize Postman collections for API testing

- Created collections for Auth, Dashboard data fetch, and System Health with relevant requests.
- Added requests for login, fetching courses, programs, and health checks.
- Set up visualizers for better response representation in Postman.

+++ backend/app/schemas/people.py
@@ -258,7 +258,7 @@ class EnrollmentCreate(EnrollmentBase):
 class EnrollmentUpdate(BaseModel):
     enrollment_status: str | None = None
     grade: str | None = None
-    grade_points: Decimal | None = None
+    grade_points: float | None = None
     credits_earned: int | None = None
     attendance_percentage: Decimal | None = None
     midterm_grade: str | None = None
@@ -274,7 +274,7 @@ class EnrollmentOut(BaseModel):
     enrollment_date: datetime | None = None
     enrollment_status: str
     grade: str | None = None
-    grade_points: Decimal | None = None
+    grade_points: float | None = None
     credits_earned: int = 0
     attendance_percentage: Decimal | None = None
     midterm_grade: str | None = None
